refactor: centralize lints in workspace config and fix clippy pedantic warnings

+++ adapters/blob-adapter-fs/tests/basic_tests.rs
@@ -2,6 +2,8 @@
 //!
 //! Tests core CRUD operations for blob storage
 
+#![allow(clippy::panic, clippy::expect_used, clippy::unwrap_used)]
+
 use cloudillo_blob_adapter_fs::BlobAdapterFs;
 use cloudillo_types::blob_adapter::{BlobAdapter, CreateBlobOptions};
 use cloudillo_types::types::TnId;
@@ -31,7 +33,7 @@ async fn test_create_and_retrieve_blob() {
 
 	// Verify blob exists via stat
 	let size = adapter.stat_blob(tn_id, file_id).await.expect("Failed to stat blob");
-	assert_eq!(size as usize, test_data.len());
+	assert_eq!(size, test_data.len() as u64);
 }
 
 #[tokio::test]
@@ -77,8 +79,8 @@ async fn test_per_tenant_isolation() {
 	let size1 = adapter.stat_blob(TnId(1), file_id).await.expect("Tenant 1 blob should exist");
 	let size2 = adapter.stat_blob(TnId(2), file_id).await.expect("Tenant 2 blob should exist");
 
-	assert_eq!(size1 as usize, data1.len());
-	assert_eq!(size2 as usize, data2.len());
+	assert_eq!(size1, data1.len() as u64);
+	assert_eq!(size2, data2.len() as u64);
 }
 
 #[tokio::test]
